rename: `admin` 패키지 이름을 `auth`로 수정
