+++ be/promotion/src/main/java/woowa/promotion/admin/auth/domain/Admin.java
@@ -1,11 +1,14 @@
-package woowa.promotion.admin.admin.domain;
+package woowa.promotion.admin.auth.domain;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import woowa.promotion.global.domain.audit.AuditingFields;
-
-import javax.persistence.*;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
