+++ be/promotion/src/main/java/woowa/promotion/global/service/AuthService.java
@@ -8,8 +8,8 @@ import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import woowa.promotion.admin.admin.domain.Admin;
-import woowa.promotion.admin.admin.infrastructure.AdminRepository;
+import woowa.promotion.admin.auth.domain.Admin;
+import woowa.promotion.admin.auth.infrastructure.AdminRepository;
 import woowa.promotion.app.member.domain.Member;
 import woowa.promotion.app.member.infrastructure.MemberRepository;
 import woowa.promotion.global.domain.jwt.JwtProvider;
